fix issues & manage rooms

+++ src/utils/FileUtil.ts
@@ -27,6 +27,9 @@ export default class FileUtil {
     }
 
     public static cellsToFile(fileName: string, cells: number[][]) {
+        console.log(cells)
+        if (!cells) return
+
         const isFileExist = fs.existsSync(fileName)
         if (isFileExist) fs.unlinkSync(fileName)
